update testing workflow
