+++ README.md
@@ -26,21 +26,14 @@ For better performence, the main part of code is written in C++.
 │   ├── get_polarization.cpp //Calculate the polarization of each perovskite unit call using LAMMPS dump file
 │   └── get_polarization_displacement.cpp //Calculate the polarization displacement from LAMMPS dump file
 └── test //files related to testing
-    ├── BTO
-    │   ├── 1.lammpstrj
-    │   ├── Main.py
-    │   └── parameter.py
-    ├── BTO_large_cell
-    │   └── head.lammpstrj
+    ├── BaTiO3
+    ├── compare.py
+    ├── com_sys.py
     ├── PIN-PMN-PT
-    │   ├── 1.lammpstrj
-    │   ├── Main.py
-    │   └── parameter.py
-    └── PSZO
-        ├── 1.lammpstrj
-        ├── Main.py
-        ├── neighbor.py
-        └── parameter.py
+    ├── PIN-PMN-PT_multi_domain
+    ├── PSZO
+    ├── README.md
+    └── run_test.sh 
 ```
 
 ## Installation
@@ -70,7 +63,6 @@ chmod +x install.sh
 ## Example
 
 ## Todo list
-0. testing case
 1. interface with Python3
 2. more easy to use
     * in get_a: support output of POSCAR format. support flexible slice of frames
