+++ test/PIN-PMN-PT_multi_domain/bec_file
@@ -1,0 +1,1 @@
+0,3.502,0,0,0,0,0,0,5.434,0,5.434,5.434,5.434,0.00,-2.97866666666667
