+++ test/PSZO/bec_file
@@ -1,0 +1,1 @@
+0,3.46,0,2.74,0,0,0,0,0,5.90,0,0,0,0,-3.045
