+++ test/PSZO/traj.lammpstrj
