+++ test/PSZO/type_map_file
@@ -1,0 +1,1 @@
+Ba,Pb,Ca,Sr,Bi,K,Na,Hf,Ti,Zr,Nb,Mg,In,Zn,O
