+++ test/README.md
@@ -1,0 +1,8 @@
+# How to contribute test cases
+
+1. make a new folder: 
+```Bash
+mkdir YOUR_FOLDER_NAME 
+```
+2. prepare the necessary input files
+3. prepare the test.sh
