+++ test/run_test.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+BINPATH=$1
+
+for dir in */; do 
+    cd $dir
+    bash test.sh $BINPATH
+    cd ..
+done
